third_party/blink/web_tests/fast/css/modify-stylesheet-minimal-recalc-style-expected.txt

PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 6
PASS internals.updateStyleAndReturnAffectedElementCount() is 6
PASS internals.updateStyleAndReturnAffectedElementCount() is 6
PASS internals.updateStyleAndReturnAffectedElementCount() is 6
PASS getComputedStyle(document.querySelector('.foo')).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelector('.baz')).color is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelector('.bar')).color is "rgb(0, 0, 0)"
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS getComputedStyle(document.querySelector('.foo')).color is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelector('.baz')).color is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelector('.bar')).color is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE

foo
foo
foo
bar
bar
bar
baz
baz
baz
